+++ include/engine/Node.h
@@ -3,7 +3,7 @@
 #include <vector>
 
 #include "Object.h"
-#include "Geometry.h"
+#include <engine/Geometry.h>
 
 class Node : public Object {
 public:
@@ -12,19 +12,19 @@ public:
     static std::shared_ptr<Node> create();
 
 
-    virtual void setZOrder(int zOrder);
-    virtual int getZOrder();
+    virtual void setZOrder(long long zOrder);
+    virtual long long getZOrder();
 
-    virtual void setScale(float scale);
-    virtual float getScale();
+    virtual void setScale(double scale);
+    virtual double getScale();
 
     virtual void setPosition(Point position);
     virtual Point getPosition() const;
-    virtual void setPosition(float x, float y);
-    virtual void setPositionX(float x);
-    virtual float getPositionX() const;
-    virtual void setPositionY(float y);
-    virtual float getPositionY() const;
+    virtual void setPosition(double x, double y);
+    virtual void setPositionX(double x);
+    virtual double getPositionX() const;
+    virtual void setPositionY(double y);
+    virtual double getPositionY() const;
 
     virtual void setAnchorPoint(Point anchorPoint);
     virtual Point getAnchorPoint() const;
@@ -35,33 +35,33 @@ public:
     virtual void setVisible(bool visible);
     virtual bool isVisible() const;
 
-    virtual void setRotation(float rotation);
-    virtual float getRotation() const;
+    virtual void setRotation(double rotation);
+    virtual double getRotation() const;
     
 
     virtual void addChild(std::shared_ptr<Node> child);
-    virtual void addChild(std::shared_ptr<Node> child, int zOrder);
-    virtual void addChild(std::shared_ptr<Node> child, int zOrder, int tag);
+    virtual void addChild(std::shared_ptr<Node> child, long long zOrder);
+    virtual void addChild(std::shared_ptr<Node> child, long long zOrder, long long tag);
     // Returns nullptr if child does not exist
-    virtual std::shared_ptr<Node> getChildByTag(int tag);
+    virtual std::shared_ptr<Node> getChildByTag(long long tag);
     virtual std::vector<std::shared_ptr<Node>> getChildren();
     
-    virtual unsigned int getChildrenCount() const;
+    virtual unsigned long long getChildrenCount() const;
     
     virtual void setParent(std::shared_ptr<Node> parent);
     virtual std::shared_ptr<Node> getParent();
     
     virtual void removeFromParent();
     virtual void removeChild(std::shared_ptr<Node> child);
-    virtual void removeChildByTag(int tag);
+    virtual void removeChildByTag(long long tag);
     virtual void removeAllChildren();
     
-    virtual void reorderChild(std::shared_ptr<Node> child, int zOrder);
+    virtual void reorderChild(std::shared_ptr<Node> child, long long zOrder);
     
     virtual void sortAllChildren();
     
     
-    virtual void setTag(int tag);
+    virtual void setTag(long long tag);
     
     virtual std::shared_ptr<void> getUserData();
     virtual void setUserData(std::shared_ptr<void> userData);
@@ -76,18 +76,18 @@ public:
 protected:
     Node();
 
-    float m_rotation;
-    float m_scale;
+    double m_rotation;
+    double m_scale;
     Point m_position;
     Point m_anchorPoint;
     Size m_contentSize;
     
-    int m_zOrder;
+    long long m_zOrder;
     
     std::vector<std::shared_ptr<Node>> m_children;
     std::shared_ptr<Node> m_parent;
     
-    int m_tag;
+    long long m_tag;
     std::shared_ptr<void> m_userData;
 
     bool m_visible;
